Ignore recreating tray after session idle
The disable() of extension is called quickly after session status
changed to idle. We make a mark when we catch this status change
signal and then ignore recreating tray in the following disable(),
and finally reset the mark to avoid normal usage of extension
enable/disable switch.

When session status changed from idle to available, e.g. unlocking
screen, this signal can be caught, but seems there is no easy way to
avoid recreating tray then. Here still does not cover such case when
the user locks the screen manually and then unlocks it.

Inspired by gnome-shell-extensions-morphine [1],
via the long discussion gnome-shell-extension-appindicator #75 [2].
Seems the better way is to connect to DBus but that's too complex.

[1]: https://github.com/JasonLG1979/gnome-shell-extensions-morphine
[2]: https://github.com/ubuntu/gnome-shell-extension-appindicator/issues/75